// problem5/express-crypto-api/src/controllers/api/product/enums/product-error.enum.ts
export enum ProductError {
    PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',

    PRODUCT_NAME_EMPTY = 'PRODUCT_NAME_EMPTY',
    PRODUCT_NAME_INVALID = 'PRODUCT_NAME_INVALID',
    PRODUCT_NAME_EXISTED = 'PRODUCT_NAME_EXISTED',

    PRODUCT_DESCRIPTION_EMPTY = 'PRODUCT_DESCRIPTION_EMPTY',
    PRODUCT_DESCRIPTION_INVALID = 'PRODUCT_DESCRIPTION_INVALID',

    PRODUCT_PRICE_EMPTY = 'PRODUCT_PRICE_EMPTY',
    PRODUCT_PRICE_INVALID = 'PRODUCT_PRICE_INVALID',

    PRODUCT_BLOCKCHAIN_EMPTY = 'PRODUCT_BLOCKCHAIN_EMPTY',
    PRODUCT_BLOCKCHAIN_INVALID = 'PRODUCT_BLOCKCHAIN_INVALID',

    PRODUCT_CONTRACT_ADDRESS_EMPTY = 'PRODUCT_CONTRACT_ADDRESS_EMPTY',
    PRODUCT_CONTRACT_ADDRESS_INVALID = 'PRODUCT_CONTRACT_ADDRESS_INVALID',

    PRODUCT_TOKEN_ID_EMPTY = 'PRODUCT_TOKEN_ID_EMPTY',
    PRODUCT_TOKEN_ID_INVALID = 'PRODUCT_TOKEN_ID_INVALID',

    PRODUCT_CATEGORY_EMPTY = 'PRODUCT_CATEGORY_EMPTY',
    PRODUCT_CATEGORY_INVALID = 'PRODUCT_CATEGORY_INVALID',

    PRODUCT_SUPPLY_EMPTY = 'PRODUCT_SUPPLY_EMPTY',
    PRODUCT_SUPPLY_INVALID = 'PRODUCT_SUPPLY_INVALID',

    PRODUCT_IS_NFT_EMPTY = 'PRODUCT_IS_NFT_EMPTY',
    PRODUCT_IS_NFT_INVALID = 'PRODUCT_IS_NFT_INVALID',

    PRODUCT_IMAGE_URL_EMPTY = 'PRODUCT_IMAGE_URL_EMPTY',
    PRODUCT_IMAGE_URL_INVALID = 'PRODUCT_IMAGE_URL_INVALID',

    PRODUCT_METADATA_TRAIT_TYPE_EMPTY = 'PRODUCT_METADATA_TRAIT_TYPE_EMPTY',
    PRODUCT_METADATA_TRAIT_TYPE_INVALID = 'PRODUCT_METADATA_TRAIT_TYPE_INVALID',
    PRODUCT_METADATA_VALUE_INVALID = 'PRODUCT_METADATA_VALUE_INVALID',
}